+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file. Based on [K
 
 ## [Unreleased]
 
+## [1.0.0] - 2021-08-21
+
 ### Added
 - Historic results in Ranking page
+- Tables for Copa championship
+- Bolao positions history in results
+
+### Changed
+- Layouts for bolao
 
 ## [0.2.0] - 2021-07-27
 
@@ -24,6 +31,7 @@ All notable changes to this project will be documented in this file. Based on [K
 - Minimal ranking page
 - Table from games from first season year 20
 
-[unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.2.0...v1.0.0
 [0.2.0]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/v0.1.0